# apps/backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Verifolio API
  description: "Verifolio backend API. Trust model: verification signals, not binary\
    \ verified flags."
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/recommender/consent:
    post:
      tags:
      - recommender-flow-controller
      operationId: consent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentDecisionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecommenderStatusResponse"
        "401":
          description: No active recommender session
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not a recommender session or missing CSRF token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Consent gate already passed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/v1/privacy/recommender-requests:
    post:
      tags:
      - recommender-dsr-controller
      operationId: intake
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommenderDataRequestRequest"
        required: true
      responses:
        "202":
          description: Accepted (no match/enumeration signal is leaked)
        "400":
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/v1/privacy/recommender-requests/{id}/verify:
    post:
      tags:
      - recommender-dsr-controller
      operationId: verify
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommenderDsrVerifyRequest"
        required: true
      responses:
        "200":
          description: Code accepted; consent withdrawal executed or request recorded
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecommenderDsrVerifyResponse"
        "400":
          description: Invalid or expired code
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Request not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Request already verified or terminal
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
        requestId:
          type:
          - string
          - "null"
...
    RecommenderStatusResponse:
      type: object
      properties:
        status:
          type: string
    ConsentDecisionRequest:
      type: object
      properties:
        accepted:
          type:
          - boolean
          - "null"
        crossBorderAccepted:
          type:
          - boolean
          - "null"
        reasonCategory:
          type:
          - string
          - "null"
          enum:
          - DONT_KNOW_REQUESTER
          - TOO_BUSY
          - NOT_COMFORTABLE
          - OTHER
      required:
      - accepted
    RecommenderDataRequestRequest:
      type: object
      properties:
        email:
          type:
          - string
          - "null"
          format: email
          maxLength: 320
          minLength: 0
        referenceRequestEmailHint:
          type:
          - string
          - "null"
          maxLength: 320
          minLength: 0
      required:
      - email
    RecommenderDsrVerifyRequest:
      type: object
      properties:
        code:
          type:
          - string
          - "null"
          minLength: 1
          pattern: "\\d{6}"
        type:
          type:
          - string
          - "null"
          enum:
          - DELETION
          - EXPORT
          - REGION_MIGRATION
          - CONSENT_WITHDRAWAL
          - CORRECTION
        referenceRequestId:
          type:
          - string
          - "null"
      required:
      - code
      - type
    RecommenderDsrVerifyResponse:
      type: object
      properties:
        status:
          type: string
        executed:
          type: boolean
        dueAt:
          type:
          - string
          - "null"
...